// src/app/shared/interface/production-value.interface.ts
export interface ProductionValueInterface {
  commissionRental: string;
  commissionSales: string;
  dealRental: string;
  dealSales: string;
  expenseRental: string;
  expenseSales: string;
  productionRental: string;
  productionSales: string;
  salaryRental: string;
  salarySales: string;
  targetRental: string;
  targetSales: string;
}